adding confirmation on delete

+++ AngularJSAuthentication.Web/app/controllers.js
@@ -19,9 +19,9 @@
         .module('VirtualClarityApp')
         .controller('usermanagementCtrl', usermanagementCtrl);
 
-    usermanagementCtrl.$inject = ['userManagementService', 'rolesService', 'UserModel', 'authService'];
+    usermanagementCtrl.$inject = ['userManagementService', 'rolesService', 'UserModel', 'authService', 'utility'];
 
-    function usermanagementCtrl(userManagementService, rolesService, UserModel, authService) {
+    function usermanagementCtrl(userManagementService, rolesService, UserModel, authService, utility) {
         var vm = this;
         vm.name = 'usermanagementCtrl';
 
@@ -29,19 +29,25 @@
         vm.newUser = new UserModel();
         vm.users = [];
 
+        vm.editUser = function(id) {
+            //todo
+        }
+
         vm.createUser = function () {
             authService.saveRegistration(vm.newUser).then(function () {
                 //todo consider sending back in payload so we dont have to do another call to refresh the screen
                 vm.getAllUsers();
                 vm.newUser = new UserModel();
             });
-        };
+        }
 
         vm.deleteUser = function (id) {
-            userManagementService.deleteUser(id).then(function() {
-                _.remove(vm.users, { 'id': id });
+            utility.confirm("Are you sure you want to delete this user?").result.then(function() {
+                userManagementService.deleteUser(id).then(function () {
+                    _.remove(vm.users, { 'id': id });
+                });
             });
-        };
+        }
 
         vm.init = function() {
             vm.getAllUsers();
@@ -49,14 +55,30 @@
             rolesService.getAllRoles().then(function (result) {
                 vm.availableRoles = result.data;
             });
-        };
+        }
 
         vm.getAllUsers = function() {
             userManagementService.getAllUsers().then(function (result) {
                 vm.users = result.data;
             });
-        };
+        }
 
         vm.init();
+    }
+}(angular));
+
+(function(angular) {
+    'use strict';
+
+    angular
+        .module('VirtualClarityApp')
+        .controller('ConfirmCtrl', ConfirmCtrl);
+
+    ConfirmCtrl.$inject = ['$modalInstance', 'message'];
+
+    function ConfirmCtrl($modalInstance, message) {
+        var vm = this;
+        vm.message = message;
+        vm.$modalInstance = $modalInstance;
     }
 }(angular));
